Fix builds with Python disabled
pyswitch_management.py (generated by SWIG) was always included in
EXTRA_DIST, but there's obviously no rule to make it when Python (and
SWIG) aren't enabled, causing the build to fail.

This patch is an attempt to correct this by bringing the switch
management Makefile.am in line with other similar ones (e.g., topology).