Fix make dist / tarball generation
A number of files were missing from the Makefile.am files, causing the
tarball built by make dist to not actually work.  The situation should
be much better now.

Note that a tarball built from a tree configured with Python enabled can
then be used to build with or without Python, but the opposite is not
true.  If the tarball was built without Python, you will NOT be able to
build a Python-enabled version from it.