Fix swig message in openflow.h
ofp_queue_get_config_reply is the one and only struct in openflow.h
that used a flexible array member rather than a zero length array
for the last member. I think this was just an oversight, and SWIG
doesn't like it, so I've converted it to a zero length array.