Make builds quieter by default
The incredible verbosity of a normal automake build is useful as an
exception, not as a rule.  Indeed, I don't think anyone watches for
warnings when building NOX because it's just too hard to see them.
So I've now made quieter builds the default.  You can get back the
noisy ones at configure time with --disable-silent-rules or at
make time with "make V=1".